+++ source/effect_codegen_spirv.cpp
@@ -594,9 +594,9 @@ private:
 	uint32_t semantic_to_location(const std::string &semantic, uint32_t max_attributes = 1)
 	{
 		if (semantic.compare(0, 5, "COLOR") == 0)
-			return std::strtoul(semantic.c_str() + 5, nullptr, 10);
+			return static_cast<uint32_t>(std::strtoul(semantic.c_str() + 5, nullptr, 10));
 		if (semantic.compare(0, 9, "SV_TARGET") == 0)
-			return std::strtoul(semantic.c_str() + 9, nullptr, 10);
+			return static_cast<uint32_t>(std::strtoul(semantic.c_str() + 9, nullptr, 10));
 
 		if (const auto it = _semantic_to_location.find(semantic);
 			it != _semantic_to_location.end())
@@ -608,7 +608,7 @@ private:
 			digit_index--;
 		digit_index++;
 
-		const uint32_t semantic_digit = std::strtoul(semantic.c_str() + digit_index, nullptr, 10);
+		const uint32_t semantic_digit = static_cast<uint32_t>(std::strtoul(semantic.c_str() + digit_index, nullptr, 10));
 		const std::string semantic_base = semantic.substr(0, digit_index);
 
 		uint32_t location = static_cast<uint32_t>(_semantic_to_location.size());
